Changed tracking positions in int to double.

+++ src/se/liu/ida/tddd78/towerdefense/interfaces/Positionable.java
@@ -8,7 +8,7 @@ import se.liu.ida.tddd78.towerdefense.objects.basic.Point;
 public interface Positionable {
     public Point getPosition();
 
-    public void setPosition(int x, int y);
+    public void setPosition(double x, double y);
 
 	public void setPosition(Point position);
 }
